The user can now run with no auth-related environment variables, in which case they will need to do auth at start and can then use Datalab and gcloud within Datalab.
Or they can set GOOGLE_APPLICATION_CREDENTIALS to point to a file (possibly
outside the container if mapped in) and then they may not need to do auth if
the file already exists (but gcloud won't work).

Or they can set CLOUDSDK_CONFIG to point to a directory (possibly outside
the container) in which case they may not need to do auth and gclould will
work (as the directory is used for both app and user credentials).

Outside of Datalab they should either use one of the above two env vars or
if they are not in Docker we can kick off an auth flow in the users local
browser.

Also in this PR:

- hiding help pane by default.
- removing some unused socket.io code
- some progress towards getting ipywidgets working
- fix 'Convert to Python' menu option
- support opening non-notebook files
- some fixes to the Sphinx docs